fix local schedule for different groups
