+++ lib/common/database/database.dart
@@ -4,19 +4,9 @@ import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as p;
+import 'package:uneconly/common/database/tables/lessons.dart';
 
 part 'database.g.dart';
-
-@DataClassName('LessonDatabaseEntity')
-class Lessons extends Table {
-  IntColumn get id => integer().autoIncrement()();
-  TextColumn get name => text()();
-  DateTimeColumn get start => dateTime()();
-  DateTimeColumn get end => dateTime()();
-  TextColumn get professor => text().nullable()();
-  TextColumn get location => text()();
-  DateTimeColumn get createdAt => dateTime()();
-}
 
 @DriftDatabase(tables: [Lessons])
 class MyDatabase extends _$MyDatabase {
@@ -26,7 +16,23 @@ class MyDatabase extends _$MyDatabase {
   // you should bump this number whenever you change or add a table definition.
   // Migrations are covered later in the documentation.
   @override
-  int get schemaVersion => 1;
+  int get schemaVersion => 2;
+
+  @override
+  MigrationStrategy get migration {
+    return MigrationStrategy(
+      onCreate: (Migrator m) async {
+        await m.createAll();
+      },
+      onUpgrade: (Migrator m, int from, int to) async {
+        if (from < 2) {
+          // we added the dueDate property in the change from version 1 to
+          // version 2
+          await m.addColumn(lessons, lessons.groupId);
+        }
+      },
+    );
+  }
 }
 
 LazyDatabase _openConnection() {
